+++ application/src/main/java/org/opentripplanner/graph_builder/module/osm/TurnRestrictionTag.java
@@ -38,5 +38,7 @@ class TurnRestrictionTag {
     RIGHT,
     U,
     STRAIGHT,
+    ENTRY,
+    EXIT,
   }
 }