+++ application/src/main/java/org/opentripplanner/transfer/regular/model/PathTransfer.java
@@ -5,7 +5,7 @@ import java.util.EnumSet;
 import java.util.List;
 import java.util.Optional;
 import org.locationtech.jts.geom.LineString;
-import org.opentripplanner.raptor.spi.RaptorCostConverter;
+import org.opentripplanner.routing.cost.CostLimit;
 import org.opentripplanner.street.geometry.GeometryUtils;
 import org.opentripplanner.street.model.StreetMode;
 import org.opentripplanner.street.model.edge.Edge;
@@ -15,10 +15,7 @@ import org.opentripplanner.street.search.state.EdgeTraverser;
 import org.opentripplanner.street.search.state.StateEditor;
 import org.opentripplanner.transfer.constrained.model.ConstrainedTransfer;
 import org.opentripplanner.transit.model.site.StopLocation;
-import org.opentripplanner.utils.logging.Throttle;
 import org.opentripplanner.utils.tostring.ToStringBuilder;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * Represents a transfer for a set of modes between stops with the street network path attached to it.
@@ -31,8 +28,6 @@ import org.slf4j.LoggerFactory;
  */
 public class PathTransfer implements Serializable {
 
-  private static final Logger LOG = LoggerFactory.getLogger(PathTransfer.class);
-
   public final StopLocation from;
 
   public final StopLocation to;
@@ -42,10 +37,6 @@ public class PathTransfer implements Serializable {
   private final List<Edge> edges;
 
   private final EnumSet<StreetMode> modes;
-
-  private static final Throttle THROTTLE_COST_EXCEEDED = Throttle.ofOneSecond();
-
-  protected static final int MAX_TRANSFER_COST = 2_000_000;
 
   public PathTransfer(
     StopLocation from,
@@ -97,12 +88,11 @@ public class PathTransfer implements Serializable {
     if (edges == null || edges.isEmpty()) {
       WalkRequest walkReq = request.walk();
       double durationSeconds = distanceMeters / walkReq.speed();
-      final double domainCost = costLimitSanityCheck(durationSeconds * walkReq.reluctance());
       return Optional.of(
         new DefaultRaptorTransfer(
           to.getIndex(),
           (int) Math.ceil(durationSeconds),
-          RaptorCostConverter.toRaptorCost(domainCost),
+          CostLimit.toRaptorCostWholeSeconds(durationSeconds * walkReq.reluctance()),
           this
         )
       );
@@ -117,7 +107,7 @@ public class PathTransfer implements Serializable {
       new DefaultRaptorTransfer(
         to.getIndex(),
         (int) s.getElapsedTimeSeconds(),
-        RaptorCostConverter.toRaptorCost(costLimitSanityCheck(s.getWeight())),
+        CostLimit.toRaptorCostWholeSeconds(s.getWeight()),
         this
       )
     );
@@ -132,34 +122,5 @@ public class PathTransfer implements Serializable {
       .addColSize("edges", edges)
       .addColSize("modes", modes)
       .toString();
-  }
-
-  /**
-   * Since transfer costs are not computed through a full A* with pruning they can incur an
-   * absurdly high cost that overflows the integer cost inside RAPTOR
-   * (https://github.com/opentripplanner/OpenTripPlanner/issues/5509).
-   * <p>
-   * An example would be a transfer using lots of stairs being used on a wheelchair when no
-   * wheelchair-specific one has been generated.
-   * (see https://docs.opentripplanner.org/en/dev-2.x/Accessibility/).
-   * <p>
-   * For this reason there is this sanity limit that makes sure that the transfer cost stays below a
-   * limit that is still very high (several days of transit-equivalent cost) but far away from the
-   * integer overflow.
-   *
-   * @see EdgeTraverser
-   * @see RaptorCostConverter
-   */
-  private int costLimitSanityCheck(double cost) {
-    if (cost >= 0 && cost <= MAX_TRANSFER_COST) {
-      return (int) cost;
-    } else {
-      THROTTLE_COST_EXCEEDED.throttle(() ->
-        LOG.warn(
-          "Transfer exceeded maximum cost. Please consider changing the transfer cost calculation. More information: https://github.com/opentripplanner/OpenTripPlanner/pull/5516#issuecomment-1819138078"
-        )
-      );
-      return MAX_TRANSFER_COST;
-    }
   }
 }