+++ application/src/test/java/org/opentripplanner/transit/service/StopTimesHelperTest.java
@@ -16,6 +16,7 @@ import org.opentripplanner.TestOtpModel;
 import org.opentripplanner.core.model.id.FeedScopedId;
 import org.opentripplanner.model.StopTimesInPattern;
 import org.opentripplanner.model.TripTimeOnDate;
+import org.opentripplanner.transit.api.request.CancellationPolicy;
 import org.opentripplanner.transit.api.request.TripTimeOnDateRequest;
 import org.opentripplanner.transit.model.network.TripPattern;
 
@@ -359,7 +360,7 @@ class StopTimesHelperTest {
       .withTime(SERVICE_DATE.atStartOfDay(transitService.getTimeZone()).toInstant())
       .withTimeWindow(Duration.ofHours(24))
       .withNumberOfDepartures(10)
-      .withIncludeCancelledTrips(true)
+      .withCancellationPolicy(CancellationPolicy.INCLUDE_CANCELLATIONS)
       .build();
 
     var result = stopTimesHelper.findTripTimesOnDate(request);